refactor:add ThemeAndLocalePreview
